+++ example/test/widget_test.dart
@@ -11,32 +11,20 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:example/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
-
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsNothing);
-
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
-
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsOneWidget);
-  });
-
-  testWidgets('Tests the revealable responses to `revealWhenEmpty=true`', (WidgetTester tester) async {
+  testWidgets('Tests the revealable responses to `revealWhenEmpty=true`',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MyApp());
     expect(find.byKey(Key('scrollable-row')), findsOneWidget);
   });
 
-  testWidgets('Tests the revealable responses to `revealWhenEmpty=false`', (WidgetTester tester) async {
+  testWidgets('Tests the revealable responses to `revealWhenEmpty=false`',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MyApp(revealWhenEmpty: false));
     expect(find.byKey(Key('scrollable-row')), findsNothing);
   });
 
-  testWidgets('Tests that dragging the list down reveals our item', (WidgetTester tester) async {
+  testWidgets('Tests that dragging the list down reveals our item',
+      (WidgetTester tester) async {
 // Build our app and trigger a frame.
     await tester.pumpWidget(MyApp(revealWhenEmpty: false));